Corrected problem where display was rotated 180 degrees.
This problem was introduced by changes to the SHARP LCD driver.  An extraneous
call to setRotation(2) is in the begin method but the logic to actually
implement screen rotations wasn't implemented.  Later it was implemented and the
setRotation(2) had the effect of rotating the display.  As a workaround, I now
call setRotation(0) just after display.begin().